\name{Cppfft4-package}
\alias{Cppfft4-package}
\alias{Cppfft4}
\docType{package}
\title{
Helper functions for numerically approximating integrodifference equations
}
\description{
This contains three functions CppNormKern(), CppReflFunc(), CppshiftmatFunc().
}
\details{
\tabular{ll}{
Package: \tab Cppfft4\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-03\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Devin W. Goodsman

Maintainer: Devin Goodsman <goodsman@ualberta.ca>
}
\references{
}
\keyword{Integrodifference equation}
\seealso{
}
\examples{
}
